INC-192 refactor: update imports
